Adds ChessColor enum
Adds an enum to store the white and black colors of chess pieces,
intended to be easier to read than passing around a boolean value to
define if a piece is white.
Also includes DRAW, to store a condition where the game ends in a draw.
This is not intended to be a valid color for a piece, possibly change
this implementation to only allow for WHITE and BlACK to be passed as
the color of a piece.