Adds Turn and DetailedTurn classes
Adds classes to describe turns taken by the player.
Turn describes only the current and proposed position of the piece
moving. This is intended to be used to pass user input to the game
processing.
DetailedTurn extends Turn but also includes all the information required
to create chess algebraic notation for a record of a game, including the
piece moving, whether the piece captured on that turn, whether the piece
put the opposing king in check, and whether the piece put the opposing
king in checkmate.