Creates PieceDetailTurn, renames DetailedTurn to FullDetailTurn
Creates a new class, PieceDetailTurn, that is between Turn and
FullDetailTurn, which includes the piece moved and captured but not
check/checkmate information. This class is intended to allow for
temporary moves that can be later undone, to check if a move puts the
same color king into check, and includes all the information necessary
to do that.
Updated the name, fields, and methods of FullDetailTurn to reflect the
change in the inheritance hierarchy and maintain name scheme.