Splits Piece into a hierarchy
Splits the Piece class into extensions of the Piece abstract class. The
StoredPiece is identical to the original Piece class, and retains the
same functionality (other than formatting changes). The QuickMovePiece
stores an existing Piece object and a new location, which allows it to
act as a Piece at the new Position without actually modifying the
original piece.