Adds Board abstract, PlayingBoard, and QuickMoveBoard
Adds class PlayingBoard to handle storage of the game state. The
abstract class Board and implementation QuickMoveBoard allow for the
creation of quickmoves, which allow a move to be tested for putting same
color pieces in check before actually allowing the move to occur.
The QuickMoveBoard takes an existing Board and a move that modifies it,
and acts like a board where that position is played.