Adds a constructor to StoredPiece that takes a position.
Allows construction of a StoredPiece with its position. This is due to changes coming to the PlayingBoard class, which will add the position at time of construction and then pass a list of all the pieces to add to the board, rather than adding each piece individually.
Also adds position to the toString method of StoredPiece to aid in debugging.