Switches PlayingBoard to store the pieces in a hashmap, removes IndexValue class.
The hashmap PlayingBoard will enable (slightly) less memory use, especially as a game progresses. The IndexValue class is no longer necessary, as there is no longer a need to convert between one and zero based indexes, as the Positions alone work as keys for the HashMap.
Adds hashCode method to Position to allow proper functionality of the HashMap.
Removes methods that facilitated interacting with the PlayingBoard 2D array, as the access and setting methods of the HashMap are harder to mess up.